+++ src/rust/bitbox02-rust/src/workflow/mnemonic.rs
@@ -65,21 +65,27 @@ pub async fn confirm_word(choices: &[&str], title: &str) -> Result<u8, CancelErr
     with_cancel("Recovery\nwords", &mut component, &result).await
 }
 
-/// Given 23 initial words, this function returns list of candidate words for the last word, such
-/// that the resulting bip39 phrase has a valid checksum. There are always exactly 8 such words.
-/// `entered_words` must contain 23 words from the BIP39 wordlist.
-fn lastword_choices(entered_words: &[&str]) -> Vec<zeroize::Zeroizing<String>> {
-    if entered_words.len() != 23 {
-        panic!("must have entered 23 words");
-    }
+/// Given 11/17/23 initial words, this function returns a list of candidate words for the last word,
+/// such that the resulting bip39 phrase has a valid checksum. There are always exactly 8 such words
+/// for 24 word mnemonics, 32 words for 18 word mnemonics and 128 words for 12 word mnemonics.
+/// `entered_words` must contain 11/17/23 words from the BIP39 wordlist.
+/// The result is the list of indices of the words in the BIP39 wordlist.
+fn lastword_choices(entered_words: &[&str]) -> Vec<u16> {
+    let (seed_len_bits, checksum_len_bits, bitmask_seed) = match entered_words.len() {
+        11 => (128, 4, 0b10000000),
+        17 => (192, 6, 0b11100000),
+        23 => (256, 8, 0b11111000),
+        _ => panic!("invalid number of entered words"),
+    };
+    let num_candidates = 1 << (seed_len_bits % 11);
 
-    // A 24 word seedphrase encodes 24*11 bits (33 bytes). The last byte is the checksum (hash over
-    // the first 32 bytes). The last word, 11 bits, is the last 3 bits of the seed plus 8 bits of
-    // the checksum. We first need the first 23 words converted to bytes so we can enumerate the 8
-    // choices for the last word. libwally only lets us convert 24 words if the checksum
-    // matches. Instead of rolling our own decoding function, we quickly find one valid word by
-    // brute-force. We need to check at most 256 words for that, as there is exactly one valid word
-    // for each 256 words block.
+    // A seedphrase for 12/18/24 words encodes 128/192/256 bits. The last 4/6/8 bits is the checksum
+    // (hash over the first 32 bytes). The last word, 11 bits, is the last 7/5/3 bits of the seed
+    // plus 4/6/8 bits of the checksum. We first need the first 11/17/23 words converted to bytes so
+    // we can enumerate the 8/32/128 choices for the last word. libwally only lets us convert a
+    // seedphrase if the checksum matches. Instead of rolling our own decoding function, we quickly
+    // find one valid word by brute-force. We need to check at most 16/64/256 words for that, as
+    // there is exactly one valid word for each 256 words block.
     let mut seed: zeroize::Zeroizing<Vec<u8>> = {
         let mut i = 0;
         loop {
@@ -92,26 +98,120 @@ fn lastword_choices(entered_words: &[&str]) -> Vec<zeroize::Zeroizing<String>> {
                 break seed;
             }
             i += 1;
-            if i >= 256 {
-                // There must be a valid word in the first 256 bip39 words. Something went wrong.
+            if i >= 2048 / num_candidates {
+                // There must be a valid word in the first 16/64/256 bip39 words. Something went
+                // wrong.
                 panic!("Could not find a valid word");
             }
         }
     };
-
-    // Generate all 8 words matching the bip39 checksum.
-    (0..8)
-        .map(|i| {
-            // Set last three bits of the seed to `i`.
-            seed[31] &= 0b11111000;
-            seed[31] |= i;
+    let seed_len = seed.len();
+    // Generate all words matching the bip39 checksum.
+    (0..num_candidates)
+        .map(|i: u16| {
+            // Set last 7/5/3 bits of the seed to `i`.
+            seed[seed_len - 1] &= bitmask_seed;
+            seed[seed_len - 1] |= i as u8;
             // Compute checksum.
             let hash = Sha256::digest(&seed);
-            // Last word is 11 bits: <last 3 bits of the seed || 8 bits checksum>.
-            let word_idx: u16 = ((i as u16) << 8) | (hash[0] as u16);
-            bitbox02::keystore::get_bip39_word(word_idx).unwrap()
+            // Last word is 11 bits: <last 7/5/3 bits of the seed || 4/6/8 bits checksum>.
+            let word_idx: u16 =
+                (i << checksum_len_bits) | (hash[0] >> (8 - checksum_len_bits)) as u16;
+            word_idx
         })
         .collect()
+}
+
+fn lastword_choices_strings(entered_words: &[&str]) -> Vec<zeroize::Zeroizing<String>> {
+    lastword_choices(entered_words)
+        .into_iter()
+        .map(|word_idx| bitbox02::keystore::get_bip39_word(word_idx).unwrap())
+        .collect()
+}
+
+/// Select the 24th word from a list of 8 valid candidate words presented as a menu.
+/// Returns `Ok(None)` if the user chooses "None of them".
+/// Returns `Ok(Some(word))` if the user chooses a word.
+/// Returns `Err(CancelError::Cancelled)` if the user cancels.
+async fn get_24th_word(
+    title: &str,
+    entered_words: &[&str],
+) -> Result<Option<zeroize::Zeroizing<String>>, CancelError> {
+    let mut choices = lastword_choices_strings(entered_words);
+    // Add one more menu entry.
+    let none_of_them_idx = {
+        choices.push(zeroize::Zeroizing::new("None of them".into()));
+        choices.len() - 1
+    };
+    loop {
+        match super::menu::pick(&as_str_vec(&choices), Some(title)).await {
+            Err(super::menu::CancelError::Cancelled) => return Err(CancelError::Cancelled),
+            Ok(choice_idx) if choice_idx as usize == none_of_them_idx => {
+                let params = confirm::Params {
+                    title: "",
+                    body: "Invalid. Check\nrecovery words.\nRestart?",
+                    ..Default::default()
+                };
+                if let Ok(()) = confirm::confirm(&params).await {
+                    return Ok(None);
+                }
+            }
+            Ok(choice_idx) => {
+                // Confirm word picked from menu again, as a typo here would be extremely annoying.
+                // Double checking is also safer, as the user might not even realize they made a typo.
+                let word = choices[choice_idx as usize].clone();
+                if let Ok(()) = confirm::confirm(&confirm::Params {
+                    title,
+                    body: &word,
+                    ..Default::default()
+                })
+                .await
+                {
+                    return Ok(Some(word));
+                }
+            }
+        }
+    }
+}
+
+/// Select the last word of a 12 or 18 word mnemonic from a list of valid candidate words. The input
+/// is the trinary input keyboard with the wordlist restricted to these candidates.
+///
+/// Returns `Ok(word)` if the user chooses a word.
+/// Returns `Err(CancelError::Cancelled)` if the user cancels.
+async fn get_12th_18th_word(
+    title: &str,
+    entered_words: &[&str],
+) -> Result<zeroize::Zeroizing<String>, CancelError> {
+    // With 12/18 words there are 128/32 candidates, so we limit the keyboard to allow entering only
+    // these.
+    loop {
+        let choices = lastword_choices(entered_words);
+        let candidates = bitbox02::keystore::get_bip39_wordlist(Some(&choices));
+        let word = trinary_input_string::enter(
+            &trinary_input_string::Params {
+                title,
+                wordlist: Some(&candidates),
+                ..Default::default()
+            },
+            trinary_input_string::CanCancel::Yes,
+            "",
+        )
+        .await
+        .map(|s| s.as_string())?;
+
+        // Confirm word picked again, as a typo here would be extremely annoying.  Double checking
+        // is also safer, as the user might not even realize they made a typo.
+        if let Ok(()) = confirm::confirm(&confirm::Params {
+            title,
+            body: &word,
+            ..Default::default()
+        })
+        .await
+        {
+            return Ok(word);
+        }
+    }
 }
 
 /// Retrieve a BIP39 mnemonic sentence of 12, 18 or 24 words from the user.
@@ -125,7 +225,7 @@ pub async fn get() -> Result<zeroize::Zeroizing<String>, CancelError> {
     status(&format!("Enter {} words", num_words), true).await;
 
     // Provide all bip39 words to restrict the keyboard entry.
-    let bip39_wordlist = bitbox02::keystore::get_bip39_wordlist().unwrap();
+    let bip39_wordlist = bitbox02::keystore::get_bip39_wordlist(None);
 
     let mut word_idx: usize = 0;
     let mut entered_words = vec![zeroize::Zeroizing::new(String::new()); num_words];
@@ -142,63 +242,35 @@ pub async fn get() -> Result<zeroize::Zeroizing<String>, CancelError> {
         // goes forward again.
         let preset = entered_words[word_idx].as_str();
 
-        let user_entry = if word_idx == 23 {
-            // For the last word, we can restrict to a subset of bip39 words that fulfil the
-            // checksum requirement. We do this only when entering 24 words, which results in a
-            // small list of 8 valid candidates.  This special case exists so that users can
-            // generate a seed using only the device and no external software, allowing seed
-            // generation via dice throws, for example.
-
-            let mut choices = lastword_choices(&as_str_vec(&entered_words[..word_idx]));
-            // Add one more menu entry.
-            let none_of_them_idx = {
-                choices.push(zeroize::Zeroizing::new("None of them".into()));
-                choices.len() - 1
-            };
-            match super::menu::pick(&as_str_vec(&choices), Some(&title)).await {
-                Err(super::menu::CancelError::Cancelled) => {
-                    Err(trinary_input_string::Error::Cancelled)
-                }
-                Ok(choice_idx) if choice_idx as usize == none_of_them_idx => {
-                    let params = confirm::Params {
-                        title: "",
-                        body: "Recovery words\ninvalid.\nRestart?",
-                        ..Default::default()
-                    };
-                    if let Ok(()) = confirm::confirm(&params).await {
-                        return Err(CancelError::Cancelled);
+        let user_entry: Result<zeroize::Zeroizing<String>, CancelError> =
+            if word_idx == num_words - 1 {
+                // For the last word, we can restrict to a subset of bip39 words that fulfil the
+                // checksum requirement. This special case exists so that users can generate a seed
+                // using only the device and no external software, allowing seed generation via dice
+                // throws, for example.
+                if num_words == 24 {
+                    // With 24 words there are only 8 valid candidates. We presnet them as a menu.
+                    match get_24th_word(&title, &as_str_vec(&entered_words[..word_idx])).await {
+                        Ok(None) => return Err(CancelError::Cancelled),
+                        Ok(Some(r)) => Ok(r),
+                        Err(e) => Err(e),
                     }
-                    continue;
+                } else {
+                    get_12th_18th_word(&title, &as_str_vec(&entered_words[..word_idx])).await
                 }
-                Ok(choice_idx) => {
-                    // Confirm word picked from menu again, as a typo here would be extremely annoying.
-                    // Double checking is also safer, as the user might not even realize they made a typo.
-                    let word = choices[choice_idx as usize].clone();
-                    match confirm::confirm(&confirm::Params {
+            } else {
+                trinary_input_string::enter(
+                    &trinary_input_string::Params {
                         title: &title,
-                        body: &word,
+                        wordlist: Some(&bip39_wordlist),
                         ..Default::default()
-                    })
-                    .await
-                    {
-                        Err(confirm::UserAbort) => continue,
-                        Ok(()) => Ok(word),
-                    }
-                }
-            }
-        } else {
-            trinary_input_string::enter(
-                &trinary_input_string::Params {
-                    title: &title,
-                    wordlist: Some(&bip39_wordlist),
-                    ..Default::default()
-                },
-                trinary_input_string::CanCancel::Yes,
-                preset,
-            )
-            .await
-            .map(|s| s.as_string())
-        };
+                    },
+                    trinary_input_string::CanCancel::Yes,
+                    preset,
+                )
+                .await
+                .map(|s| s.as_string())
+            };
 
         match user_entry {
             Err(CancelError::Cancelled) => {
@@ -278,19 +350,80 @@ mod tests {
 
     #[test]
     fn test_lastword_choices() {
+        // 23 words
+
         assert_eq!(
             &as_str_vec(&bruteforce_lastword(&["violin"; 23])),
             &["boss", "coyote", "dry", "habit", "panel", "regular", "speed", "winter"]
         );
 
         assert_eq!(
-            &lastword_choices(&["violin"; 23]),
+            &lastword_choices_strings(&["violin"; 23]),
             &bruteforce_lastword(&["violin"; 23]),
         );
 
         let mnemonic = "side stuff card razor rescue enhance risk exchange ozone render large describe gas juice offer permit vendor custom forget lecture divide junior narrow".split(' ').collect::<Vec<&str>>();
         assert_eq!(
-            &lastword_choices(&mnemonic),
+            &lastword_choices_strings(&mnemonic),
+            &bruteforce_lastword(&mnemonic)
+        );
+
+        // 17 words
+
+        assert_eq!(
+            &as_str_vec(&bruteforce_lastword(&["violin"; 17])),
+            &[
+                "all", "appear", "bike", "book", "cash", "click", "cycle", "disagree", "donate",
+                "essence", "fence", "gadget", "ghost", "hotel", "industry", "lab", "lizard",
+                "modify", "much", "oblige", "pond", "pull", "raccoon", "reunion", "side", "smoke",
+                "steak", "taxi", "tongue", "used", "wall", "wonder"
+            ]
+        );
+
+        assert_eq!(
+            &lastword_choices_strings(&["violin"; 17]),
+            &bruteforce_lastword(&["violin"; 17]),
+        );
+
+        let mnemonic = "alpha write diary chicken cable spoil dirt hair bike fiction system bright mimic garage giggle involve leisure".split(' ').collect::<Vec<&str>>();
+        assert_eq!(
+            &lastword_choices_strings(&mnemonic),
+            &bruteforce_lastword(&mnemonic)
+        );
+
+        // 11 words
+
+        assert_eq!(
+            &as_str_vec(&bruteforce_lastword(&["violin"; 11])),
+            &[
+                "achieve", "actress", "affair", "all", "amount", "arm", "arrest", "attend",
+                "bacon", "bar", "best", "bitter", "body", "box", "brush", "bulk", "cage", "carry",
+                "chalk", "chicken", "city", "climb", "color", "convince", "cotton", "crawl",
+                "cruel", "dawn", "degree", "desk", "diet", "disease", "double", "dumb", "duty",
+                "elder", "enemy", "engage", "essay", "evoke", "faint", "family", "feel", "finger",
+                "flush", "foil", "frame", "garage", "giant", "glue", "gorilla", "green", "habit",
+                "health", "horse", "hover", "illness", "inherit", "intact", "island", "keen",
+                "know", "ladder", "lawsuit", "lesson", "lobster", "love", "main", "matter",
+                "mention", "milk", "monitor", "mother", "myself", "nest", "nose", "offer", "open",
+                "outer", "paddle", "peanut", "pear", "piece", "polar", "post", "print", "pulse",
+                "purpose", "rally", "rebuild", "regret", "report", "rifle", "rocket", "royal",
+                "salon", "sea", "segment", "shallow", "ship", "similar", "slice", "snake", "soft",
+                "source", "spray", "steel", "style", "super", "swim", "talk", "tent", "they",
+                "tiny", "tone", "treat", "trim", "turtle", "unaware", "upper", "van", "viable",
+                "vivid", "walnut", "weird", "window", "worth", "zero"
+            ]
+        );
+
+        assert_eq!(
+            &lastword_choices_strings(&["violin"; 11]),
+            &bruteforce_lastword(&["violin"; 11]),
+        );
+
+        let mnemonic = "outer elite desert faint cliff useless teach screen combine exercise below"
+            .split(' ')
+            .collect::<Vec<&str>>();
+        assert_eq!(
+            &lastword_choices_strings(&mnemonic),
             &bruteforce_lastword(&mnemonic)
         );
     }