+++ CHANGELOG.md
@@ -10,6 +10,9 @@ customers cannot upgrade their bootloader, its changes are recorded separately.
 - Disable screensaver when displaying a receive address, confirming a transaction, and other interactive actions
 - Add Sepolia testnet for Ethereum
 - Added a disclaimer screen before the recovery words are displayed
+- Verifiable seed generation: when restoring from 12 or 18 recovery words, for the final word, restrict input to the valid candidate words which result in a valid checksum.
+  For 24 words, this feature was already introduced in 9.4.0.
+
 
 ### 9.15.0
 - Security bugfix: check index of an input's previous output to prevent the fee attack originally